+++ src/lib.rs
@@ -1,5 +1,6 @@
 #![deny(clippy::all)]
 
+pub mod csv;
 pub mod error;
 pub mod model;
 pub mod reader;